customer complete

+++ CaseMngmt.Server/Controllers/CustomerController.cs
@@ -1,13 +1,11 @@
 ﻿using CaseMngmt.Models.Customers;
-using CaseMngmt.Service;
 using CaseMngmt.Service.Customers;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace CaseMngmt.Server.Controllers
 {
     [ApiController]
-    [Route("api/[controller]/[action]")]
+    [Route("api/[controller]")]
     public class CustomerController : ControllerBase
     {
         private readonly ILogger<CustomerController> _logger;
@@ -19,11 +17,11 @@ namespace CaseMngmt.Server.Controllers
         }
 
         [HttpGet, Route("getAll")]
-        public async Task<IActionResult> GetAll(string customerName, string phoneNumber, int pageSize = 25, int pageNumber = 1)
+        public async Task<IActionResult> GetAll(string? customerName = null, string? phoneNumber = null, int? pageSize = 25, int? pageNumber = 1)
         {
             try
             {
-                var result = await _service.GetAllCustomersAsync(customerName, phoneNumber, pageSize, pageNumber);
+                var result = await _service.GetAllCustomersAsync(customerName, phoneNumber, pageSize.Value, pageNumber.Value);
                 return Ok(result);
             }
             catch (Exception e)
@@ -34,16 +32,16 @@ namespace CaseMngmt.Server.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> Details(int? id)
+        public async Task<IActionResult> Details(Guid id)
         {
-            if (id == null)
+            if (id == Guid.Empty)
             {
                 return BadRequest(ModelState);
             }
 
             try
             {
-                CustomerViewModel result = await _service.GetByIdAsync(id.Value);
+                CustomerViewModel result = await _service.GetByIdAsync(id);
 
                 if (result == null)
                 {
@@ -58,45 +56,26 @@ namespace CaseMngmt.Server.Controllers
             }
         }
 
-        [HttpPost, Route("create")]
-        public async Task<IActionResult> Create(CustomerViewModel customer)
+        [HttpPost]
+        public async Task<IActionResult> Create(CustomerRequest customer)
         {
-            try
+            if (!ModelState.IsValid || customer == null)
             {
-                if (ModelState.IsValid && customer != null)
-                {
-                    var isExist = await _service.CheckCustomerExistsAsync(customer.Name);
-                    if (isExist)
-                    {
-                        return BadRequest("Customer already exists");
-                    }
-                    // TODO
-                    customer.CreatedBy = Guid.Empty;
-                    customer.CreatedDate = DateTime.UtcNow;
-
-                    var result = await _service.AddCustomerAsync(customer);
-                    return Ok(result);
-                }
-
                 return BadRequest(ModelState);
             }
-            catch (Exception e)
-            {
-                _logger.LogError(e.Message, nameof(CustomerController), true, e);
-                return BadRequest();
-            }
-        }
 
-        [HttpPut, Route("update/{newsId}")]
-        public async Task<IActionResult> Update(Guid newsId, CustomerViewModel model)
-        {
             try
             {
+                var isExist = await _service.CheckCustomerExistsAsync(customer.Name);
+                if (isExist)
+                {
+                    return BadRequest("Customer already exists");
+                }
                 // TODO
-                model.UpdatedBy = Guid.Empty;
-                model.UpdatedDate = DateTime.UtcNow;
+                // customer.CreatedBy = Guid.Empty;
 
-                var result = await _service.UpdateCustomerAsync(model);
+                var result = await _service.AddCustomerAsync(customer);
+
                 return result > 0 ? Ok(result) : BadRequest();
             }
             catch (Exception e)
@@ -106,9 +85,43 @@ namespace CaseMngmt.Server.Controllers
             }
         }
 
-        [HttpDelete, Route("delete/{id}")]
-        public async Task<IActionResult> Delete(int id)
+        [HttpPut, Route("{Id}")]
+        public async Task<IActionResult> Update(Guid Id, CustomerRequest model)
         {
+            if (Id == Guid.Empty)
+            {
+                return BadRequest();
+            }
+
+            try
+            {
+                var isExist = await _service.CheckCustomerExistsAsync(model.Name);
+                if (isExist)
+                {
+                    return BadRequest("Customer already exists");
+                }
+
+                // TODO
+                // model.UpdatedBy = Guid.Empty;
+
+                var result = await _service.UpdateCustomerAsync(Id, model);
+                return result > 0 ? Ok(result) : BadRequest();
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e.Message, nameof(CustomerController), true, e);
+                return BadRequest();
+            }
+        }
+
+        [HttpDelete, Route("{id}")]
+        public async Task<IActionResult> Delete(Guid id)
+        {
+            if (id == Guid.Empty)
+            {
+                return BadRequest(ModelState);
+            }
+
             try
             {
                 var result = await _service.DeleteAsync(id);
